+++ gulpfile.js
@@ -38,7 +38,8 @@ var templates = {
       'aarhus-second-sensor',
       'aarhus-second-sensor-single',
       'aarhus-second-countdown',
-      'aarhus-second-data'
+      'aarhus-second-data',
+      'aarhus-second-sensor-history'
     ]
   }
 };